+++ Packages/com.unity.inputsystem/InputSystem/Controls/TouchControl.cs
@@ -76,13 +76,13 @@ namespace UnityEngine.Experimental.Input.Controls
 
         public override unsafe TouchState ReadUnprocessedValueFromState(void* statePtr)
         {
-            var valuePtr = (TouchState*)(byte*)statePtr + (int)m_StateBlock.byteOffset;
+            var valuePtr = (TouchState*)((byte*)statePtr + (int)m_StateBlock.byteOffset);
             return *valuePtr;
         }
 
         public override unsafe void WriteValueIntoState(TouchState value, void* statePtr)
         {
-            var valuePtr = (TouchState*)(byte*)statePtr + (int)m_StateBlock.byteOffset;
+            var valuePtr = (TouchState*)((byte*)statePtr + (int)m_StateBlock.byteOffset);
             UnsafeUtility.MemCpy(valuePtr, UnsafeUtility.AddressOf(ref value), UnsafeUtility.SizeOf<TouchState>());
         }
     }