+++ Packages/com.unity.inputsystem/InputSystem/Plugins/EnhancedTouch/TouchSwipe.cs
@@ -1,6 +1,0 @@
-namespace UnityEngine.Experimental.Input.Touch
-{
-    public class TouchSwipe
-    {
-    }
-}