+++ Packages/com.unity.inputsystem/InputSystem/Plugins/EnhancedTouch/TouchTap.cs
@@ -1,7 +1,0 @@
-namespace UnityEngine.Experimental.Input.Touch
-{
-    [AddComponentMenu("Input/Touch/Tap")]
-    public class TouchTap : MonoBehaviour
-    {
-    }
-}